+++ tests/test_decode.py
@@ -342,37 +342,33 @@ class TestInvalidValues(unittest.TestCase):
 
 
         fixture = self.my_arff.format(data="a','c d'")
-        with self.assertRaisesRegexp(ValueError,
-                                     "Expected comma or whitespace at position "
-                                     "4, not c for line a','c d'."):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     "','c d'"):
             arff.load(fixture)
 
         fixture = self.my_arff.format(data="a b, 'c d'")
-        with self.assertRaisesRegexp(arff.BadStringValue,
-                                     "Invalid string value at line 8."):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     "'c d"):
             print(arff.load(fixture))
 
         fixture = self.my_arff.format(data="'a b', c d")
-        with self.assertRaisesRegexp(arff.BadNominalFormatting,
-                                     'Nominal data value "c d" not properly '
-                                     'quoted in line 8'):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     'c d'):
             print(arff.load(fixture))
 
     def test_sparse(self):
 
         fixture = self.my_arff.format(data="{0 a',1 'c d'}")
-        with self.assertRaisesRegexp(ValueError,
-                                     "Expected comma or whitespace at position"
-                                     " 9, not c for line {0 a',1 'c d'}."):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     "',1 'c d'\}."):
             arff.load(fixture)
 
         fixture = self.my_arff.format(data="{0 a b,1 'c d'}")
-        with self.assertRaisesRegexp(arff.BadStringValue,
-                                     "Invalid string value at line 8."):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     "b,1 'c d'"):
             print(arff.load(fixture))
 
         fixture = self.my_arff.format(data="{0 'a b', 1 c d}")
-        with self.assertRaisesRegexp(arff.BadNominalFormatting,
-                                     'Nominal data value "c d" not properly '
-                                     'quoted in line 8'):
+        with self.assertRaisesRegexp(arff.ArffException,
+                                     r'.*d\}'):
             print(arff.load(fixture))