+++ CHANGES.rst
@@ -1,6 +1,12 @@
 What's New on LIAC-ARFF
 =======================
 
+LIAC-ARFF 2.3
+
+- enh: improvements to loading runtime (issue #76)
+- fix: several bugs in decoding and encoding quoted and escaped values,
+  particularly in loading sparse ARFF.
+
 LIAC-ARFF 2.2.3
 
 - new: test for python3.7 and pypy3